backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
pillow==10.1.0
python-dotenv==1.0.0
ultralytics==8.3.30        # for YOLO drop-in later (safe to install now)
numpy<2
opencv-python-headless==4.8.1.78
mediapipe==0.10.14
tensorflow==2.13.0    # (or your trained model’s version)

typing-extensions==4.5.0
google-generativeai==0.4.0
pydantic==1.10.13
tensorflow-macos==2.13.0

